backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
supabase  # <-- Corrected name
python-jose[cryptography] # If you're using this for JWTs as discussed